Extracted NetworkMonitor
